matcher/pyproject.toml
[project]
name = "matcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "numpy>=2.3.2",
    "python-multipart>=0.0.20",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
]
